+++ src/api/types/Vec3.ts
@@ -1,0 +1,7 @@
+//  This file was auto-generated from our API Definition.
+
+export interface Vec3 {
+    x?: number | undefined;
+    y?: number | undefined;
+    z?: number | undefined;
+}